created a new Scene called ShipSetup
scene has a very simple ship move controller as well as two stars and a black hole fx i made while i was procrastonating doing my HW.  I think the lighting looks pretty good tho.  The ship controls are pretty simple, I used Odin's serializer to serialize the thrusters as interfaces so I could try a couple variations.  
WASD is forward and lateral thrusters and QE are angular thrusters


The thruster particles are for testing the visuals in a scene.

 shield shader is a WIP. I used this tutorial to make it: https://www.youtube.com/watch?v=hTJqo1HeEOs&feature=emb_logo